telegram-bot-chef/requirements.txt
aiogram==3.10.0
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
python-dotenv==1.0.0
aiohttp==3.9.1
gigachat==0.1.10
python-multipart==0.0.6